create base io file
